Separated the class list to a different file, and updated the charClass object to have more properties
